requirements.txt
Django==3.2
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.48
Pillow==9.1.1
psycopg2-binary==2.9.3
